Drobne poprawki i walidacja danych
