+++ project/app.py
@@ -1,28 +1,18 @@
 import imghdr
 import os
 import flask
-import logging
 import flask_cors
 from werkzeug.utils import secure_filename
-from werkzeug.exceptions import HTTPException
+import project.functions as f
 
 UPLOAD_FOLDER = 'uploads'
-ALLOWED_EXTENSIONS = ['.txt', '.pdf', '.png', '.jpg', '.jpeg', '.gif']
+ALLOWED_EXTENSIONS = ['.png', '.jpg', '.jpeg']
 TEMPLATE_FOLDER = os.path.abspath('../client')
 
 app = flask.Flask(__name__, template_folder=TEMPLATE_FOLDER)
 app.config["UPLOAD_PATH"] = UPLOAD_FOLDER
 app.config["UPLOAD_EXTENSIONS"] = ALLOWED_EXTENSIONS
 cors = flask_cors.CORS(app)
-
-
-def validate_image(stream):
-    header = stream.read(512)  # 512 bytes should be enough for a header check
-    stream.seek(0)  # reset stream pointer
-    format = imghdr.what(None, header)
-    if not format:
-        return None
-    return '.' + (format if format != '.jpeg' else '.jpg')
 
 
 class InvalidUsage(Exception):
@@ -48,7 +38,7 @@ def handle_invalid_usage(error):
 
 @app.route("/")
 def index():
-    files = os.listdir(app.config['UPLOAD_PATH'])
+    files = f.list_dir(app.config['UPLOAD_PATH'], app.config["UPLOAD_EXTENSIONS"])
 
     return flask.render_template('upload.html', files=files)
 
@@ -58,9 +48,9 @@ def upload_files():
     uploaded_file = flask.request.files['file']
     filename = secure_filename(uploaded_file.filename)
     if filename != '':
-        file_ext = os.path.splitext(filename)[1]
+        file_ext = (os.path.splitext(filename)[1]).lower()
 
-        if file_ext not in app.config['UPLOAD_EXTENSIONS'] or file_ext != validate_image(uploaded_file.stream):
+        if file_ext not in app.config['UPLOAD_EXTENSIONS']:
             raise InvalidUsage("Please provide valid image to upload", status_code=400)
 
         path = os.path.join(app.config['UPLOAD_PATH'], filename)
